# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API сервиса iAmOkey",
        "description": "Документация сервиса.",
        "contact": {
            "email": "nfs2025@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/comment/": {
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Создать новый комментарий",
                "description": "Создатьновый комментарий, возвращается объект комментария",
                "operationId": "30116951de0d13d6d18ca00c919259ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comments"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comments"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/comment/{id}": {
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Удалить комментарий",
                "description": "Удалить комментарий",
                "operationId": "c62e0a86bead5c4f9817a8103b523015",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/device/": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Дата регистрации устройств",
                "description": "Возвращается объект всех устройств",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Device"
                ],
                "summary": "Зарегистрировать устройства",
                "description": "Возвращается объект устройства",
                "operationId": "e2b25eda56e5eb9bfa463c751d7c61bf",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "ID устройства",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/device/{device_id}": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Дата регистрации устройства",
                "description": "Возвращается объект устройства",
                "operationId": "4b91106c4aee75c92573f50593605765",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID устройства",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/events/{patient_id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Уведомления пациента",
                "description": "Возвращается объект уведомлений пациента",
                "operationId": "3068dc5ab1616926e3ba3669a9f498b0",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "patient_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/events/": {
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Запись уведомления",
                "description": "Создать нового пациента, возвращается объект пациента",
                "operationId": "284454e652120d5dc433cfb2720d18f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Events"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Events"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/events/{id}": {
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Удалить уведомление",
                "description": "Удалить уведомление",
                "operationId": "2baad6a2e5111d5bec978166d5a89582",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id уведомления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comments": {
                "title": "Comments",
                "description": "Comments model",
                "properties": {
                    "patient_id": {
                        "title": "patient_id",
                        "description": "ID пациента",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "ID пользвателя",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "comment": {
                        "title": "comment",
                        "description": "Комментарий",
                        "type": "integer",
                        "example": "Комментарий"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Comments"
                }
            },
            "Events": {
                "title": "Events",
                "description": "Events model",
                "properties": {
                    "patient_id": {
                        "title": "Patient_id",
                        "description": "Id пациента",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "indicator_id": {
                        "title": "Indicator_id",
                        "description": "Id показаний",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "event_type": {
                        "title": "Event_type",
                        "description": "Тип события",
                        "type": "string",
                        "example": "1"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Пол",
                        "type": "integer",
                        "example": "Текстовое описание"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Events"
                }
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}